Move role getter to utils.
