+++ src/engine/core/ai/camp/camp_utils.test.ts
@@ -1,9 +1,12 @@
 import { describe, expect, it, jest } from "@jest/globals";
 
-import { startPlayingGuitar, startPlayingHarmonica } from "@/engine/core/ai/camp/camp_utils";
-import { CampManager } from "@/engine/core/ai/camp/index";
+import { getObjectCampActivityRole, startPlayingGuitar, startPlayingHarmonica } from "@/engine/core/ai/camp/camp_utils";
+import { CampManager, EObjectCampActivity, EObjectCampRole } from "@/engine/core/ai/camp/index";
+import { EStalkerState } from "@/engine/core/animation/types";
 import { IRegistryObjectState, registerCampZone, registerObject, registry } from "@/engine/core/database";
-import { GameObject } from "@/engine/lib/types";
+import { IAnimpointActionDescriptor, ISchemeAnimpointState } from "@/engine/core/schemes/stalker/animpoint";
+import { EScheme, GameObject } from "@/engine/lib/types";
+import { mockSchemeState } from "@/fixtures/engine";
 import { MockGameObject, MockIniFile } from "@/fixtures/xray";
 
 describe("camp utils", () => {
@@ -74,4 +77,70 @@ describe("camp utils", () => {
   it.todo("canPlayCampGuitar should correctly check");
 
   it.todo("canPlayCampHarmonica should correctly check");
+
+  it("getObjectCampActivityRole should correctly get object roles for generic activities", () => {
+    const object: GameObject = MockGameObject.mock();
+    const state: IRegistryObjectState = registerObject(object);
+
+    state.activeScheme = EScheme.ANIMPOINT;
+
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.IDLE)).toBe(EObjectCampRole.NONE);
+    expect(getObjectCampActivityRole(object.id(), "unexpected" as unknown as EObjectCampActivity)).toBe(
+      EObjectCampRole.NONE
+    );
+
+    state[EScheme.ANIMPOINT] = mockSchemeState(EScheme.ANIMPOINT);
+
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.IDLE)).toBe(EObjectCampRole.LISTENER);
+    expect(getObjectCampActivityRole(object.id(), "unexpected" as unknown as EObjectCampActivity)).toBe(
+      EObjectCampRole.NONE
+    );
+  });
+
+  it("getObjectCampActivityRole should correctly get object roles for harmonica/guitar", () => {
+    const object: GameObject = MockGameObject.mock();
+    const state: IRegistryObjectState = registerObject(object);
+
+    state.activeScheme = EScheme.ANIMPOINT;
+
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.HARMONICA)).toBe(EObjectCampRole.NONE);
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.GUITAR)).toBe(EObjectCampRole.NONE);
+
+    state[EScheme.ANIMPOINT] = mockSchemeState<ISchemeAnimpointState>(EScheme.ANIMPOINT, {
+      description: EStalkerState.ANIMPOINT_SIT_ASS,
+      approvedActions: $fromArray<IAnimpointActionDescriptor>([
+        { name: EStalkerState.ANIMPOINT_SIT_ASS_GUITAR, predicate: () => true },
+        { name: EStalkerState.ANIMPOINT_SIT_ASS_HARMONICA, predicate: () => true },
+      ]),
+    });
+
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.GUITAR)).toBe(EObjectCampRole.DIRECTOR);
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.HARMONICA)).toBe(EObjectCampRole.DIRECTOR);
+  });
+
+  it("getObjectCampActivityRole should correctly get object roles for story teller", () => {
+    const object: GameObject = MockGameObject.mock();
+    const state: IRegistryObjectState = registerObject(object);
+
+    state.activeScheme = EScheme.ANIMPOINT;
+
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.HARMONICA)).toBe(EObjectCampRole.NONE);
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.GUITAR)).toBe(EObjectCampRole.NONE);
+
+    state[EScheme.ANIMPOINT] = mockSchemeState<ISchemeAnimpointState>(EScheme.ANIMPOINT, {
+      approvedActions: $fromArray<IAnimpointActionDescriptor>([
+        { name: EStalkerState.ANIMPOINT_STAY_TABLE_WEAPON, predicate: () => true },
+        { name: EStalkerState.ANIMPOINT_STAY_TABLE, predicate: () => true },
+      ]),
+    });
+
+    (state[EScheme.ANIMPOINT] as ISchemeAnimpointState).description = EStalkerState.STOOP_NO_WEAP;
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.STORY)).toBe(EObjectCampRole.LISTENER);
+
+    (state[EScheme.ANIMPOINT] as ISchemeAnimpointState).description = EStalkerState.ANIMPOINT_STAY_TABLE;
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.STORY)).toBe(EObjectCampRole.DIRECTOR);
+
+    (state[EScheme.ANIMPOINT] as ISchemeAnimpointState).description = EStalkerState.ANIMPOINT_STAY_TABLE_WEAPON;
+    expect(getObjectCampActivityRole(object.id(), EObjectCampActivity.STORY)).toBe(EObjectCampRole.DIRECTOR);
+  });
 });
